refactor: misc into shared
