+++ apps/blog/src/app/components/pages/[tag]/components/NoTagsFound.tsx
@@ -3,7 +3,7 @@
 import Link from "next/link";
 import { motion } from "framer-motion";
 
-import { H2 } from "~/app/components/pages/[post]/components/mdx-custom/headers";
+import { H2 } from "~/app/components/shared/mdx-custom/headers";
 
 interface NoTagsFoundProps {
   validTags: Set<string>;
