+++ apps/blog/src/app/components/shared/mdx-custom/featured/blog/framer-motion/components/YeetMe.tsx
