+++ apps/blog/src/app/components/shared/mdx-custom/featured/blog/gsap/components/index.ts
