+++ apps/blog/src/app/components/shared/views/index.tsx
